# swagger/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.2"
  title: Simple Debts API
# during dev, should point to your local machine
host: simple-debts.herokuapp.com
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  jwt:
    description: login, get your token and put it here in format 'bearer *token*'
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /users:
    x-swagger-router-controller: usersController
    get:
      tags:
        - users
      description: Search users by name
      operationId: getUsersArrayByName
      parameters:
        - in: query
          name: name
          required: true
          type: string
          description: User name
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/User"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - jwt: []
    patch:
      tags:
        - users
      description: Updates user's profile data
      operationId: updateUserData
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: image
          type: file
          description: User's avatar. Generated on server if not set
        - in: formData
          name: name
          type: string
          required: true
          description: User's name
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/User"
        "400":
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - jwt: []
...
definitions:
...
  User:
    type: object
    properties:
      id:
        type: string
        format: id
        description: User's Id
      name:
        type: string
      picture:
        type: string
        format: url
        description: Url of user's avatar
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string